move goto_instruction_code.h to goto-programs/
The classes in goto_instruction_code.h are the representation of the
instructions of goto programs, and thus belong into the goto-programs
subdirectory.
