+++ jbmc/src/java_bytecode/code_with_references.cpp
@@ -9,8 +9,9 @@ Author: Romain Brenguier, romain.brenguier@diffblue.com
 #include "code_with_references.h"
 #include "java_types.h"
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/arith_tools.h>
-#include <util/goto_instruction_code.h>
 
 codet allocate_array(
   const exprt &expr,
