+++ jbmc/src/java_bytecode/java_bytecode_typecheck_code.cpp
@@ -11,7 +11,8 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "java_bytecode_typecheck.h"
 
-#include <util/goto_instruction_code.h>
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/std_code.h>
 
 void java_bytecode_typecheckt::typecheck_code(codet &code)
