+++ jbmc/src/java_bytecode/lift_clinit_calls.cpp
@@ -12,8 +12,9 @@ Author: Diffblue Ltd.
 
 #include <java_bytecode/java_static_initializers.h>
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/expr_iterator.h>
-#include <util/goto_instruction_code.h>
 
 #include <algorithm>
