+++ jbmc/unit/java-testing-utils/require_goto_statements.h
@@ -9,7 +9,8 @@ Author: Diffblue Ltd.
 /// \file
 /// Utilties for inspecting goto functions
 
-#include <util/goto_instruction_code.h>
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/optional.h>
 
 #include <regex>
