+++ jbmc/unit/java_bytecode/java_static_initializers/java_static_initializers.cpp
@@ -15,8 +15,9 @@ Author: Diffblue Ltd.
 #include <testing-utils/expr_query.h>
 #include <testing-utils/use_catch.h>
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/arith_tools.h>
-#include <util/goto_instruction_code.h>
 #include <util/json.h>
 #include <util/symbol_table.h>
