+++ jbmc/unit/java_bytecode/java_static_initializers/module_dependencies.txt
@@ -1,3 +1,4 @@
+goto-programs
 java_bytecode
 testing-utils
 util
