+++ src/ansi-c/expr2c_class.h
@@ -16,9 +16,10 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <unordered_map>
 #include <unordered_set>
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/bitvector_expr.h>
 #include <util/byte_operators.h>
-#include <util/goto_instruction_code.h>
 #include <util/mathematical_expr.h>
 #include <util/std_code.h>
