+++ src/cpp/cpp_typecheck_initializer.cpp
@@ -11,10 +11,11 @@ Author: Daniel Kroening, kroening@cs.cmu.edu
 
 #include "cpp_typecheck.h"
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/arith_tools.h>
 #include <util/c_types.h>
 #include <util/expr_initializer.h>
-#include <util/goto_instruction_code.h>
 #include <util/pointer_expr.h>
 #include <util/pointer_offset_size.h>
