+++ src/goto-programs/wp.cpp
@@ -13,11 +13,11 @@ Author: Daniel Kroening, kroening@kroening.com
 
 // #include <langapi/language_util.h>
 
-#include <util/goto_instruction_code.h>
-#include <util/pointer_expr.h>
-#include <util/std_code.h>
+#include "goto_instruction_code.h"
 
 #include <util/invariant.h>
+#include <util/pointer_expr.h>
+#include <util/std_code.h>
 
 bool has_nondet(const exprt &dest)
 {
