+++ src/statement-list/statement_list_typecheck.cpp
@@ -12,8 +12,9 @@ Author: Matthias Weiss, matthias.weiss@diffblue.com
 #include "statement_list_typecheck.h"
 #include "converters/statement_list_types.h"
 
+#include <goto-programs/goto_instruction_code.h>
+
 #include <util/cprover_prefix.h>
-#include <util/goto_instruction_code.h>
 #include <util/message.h>
 #include <util/namespace.h>
 #include <util/pointer_expr.h>
