+++ src/util/std_code.cpp
@@ -13,7 +13,6 @@ Author: Daniel Kroening, kroening@kroening.com
 
 #include "arith_tools.h"
 #include "c_types.h"
-#include "goto_instruction_code.h"
 #include "pointer_expr.h"
 #include "string_constant.h"
 
@@ -154,62 +153,6 @@ void code_function_bodyt::set_parameter_identifiers(
     sub.push_back(irept(ID_parameter));
     sub.back().set(ID_identifier, id);
   }
-}
-
-code_inputt::code_inputt(
-  std::vector<exprt> arguments,
-  optionalt<source_locationt> location)
-  : codet{ID_input, std::move(arguments)}
-{
-  if(location)
-    add_source_location() = std::move(*location);
-  check(*this, validation_modet::INVARIANT);
-}
-
-code_inputt::code_inputt(
-  const irep_idt &description,
-  exprt expression,
-  optionalt<source_locationt> location)
-  : code_inputt{{address_of_exprt(index_exprt(
-                   string_constantt(description),
-                   from_integer(0, index_type()))),
-                 std::move(expression)},
-                std::move(location)}
-{
-}
-
-void code_inputt::check(const codet &code, const validation_modet vm)
-{
-  DATA_CHECK(
-    vm, code.operands().size() >= 2, "input must have at least two operands");
-}
-
-code_outputt::code_outputt(
-  std::vector<exprt> arguments,
-  optionalt<source_locationt> location)
-  : codet{ID_output, std::move(arguments)}
-{
-  if(location)
-    add_source_location() = std::move(*location);
-  check(*this, validation_modet::INVARIANT);
-}
-
-code_outputt::code_outputt(
-  const irep_idt &description,
-  exprt expression,
-  optionalt<source_locationt> location)
-  : code_outputt{{address_of_exprt(index_exprt(
-                    string_constantt(description),
-                    from_integer(0, index_type()))),
-                  std::move(expression)},
-                 std::move(location)}
-{
-}
-
-void code_outputt::check(const codet &code, const validation_modet vm)
-{
-  DATA_CHECK(
-    vm, code.operands().size() >= 2, "output must have at least two operands");
 }
 
 code_fort code_fort::from_index_bounds(
